Extract business logic from controller into service
